Refactor: adjust import paths to use "util" directory instead of "utils"
